refactor: extract shared PersonSearchCombobox component
Made-with: Cursor
